Extract oversize content issue helpers partial
